requirements.txt
chatterbot
chatterbot_corpus
pywhatkit
googlesearch-python
matplotlib